skills/ketchup-brand/requirements.txt
# Ketchup Brand Dependencies
playwright>=1.40.0
colorthief>=0.2.1
Pillow>=10.0.0
requests>=2.31.0
